backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "flask-openapi3[rapidoc]>=4.0.3",
    "flask>=3.1.0",
    "python-dotenv>=1.0.1",
    "email-validator>=2.2.0",
    "flask-cors>=5.0.0",
    "faker>=33.1.0",
    "tqdm>=4.67.1",
    "flask-bcrypt>=1.0.1",
    "flask-jwt-extended>=4.7.1",
    "pymongo>=4.10.1",
    "gunicorn>=23.0.0",
]

[tool.mypy]
plugins = ['pydantic.mypy']

[tool.ruff]
cache-dir = ".venv/cache/ruff"

[tool.ruff.lint]
select = ['ALL']
ignore = [
    'COM812', # missing-trailing-comma, conflicts with formatter
    'D203',   # incorrect-blank-line-before-class, conflicts with no-blank-line-before-class
    'D213',   # multi-line-summary-second-line, conflicts with multi-line-summary-first-line
    'S606',   # start-process-with-no-shell, conflicts with start-process-with-a-shell
]

[tool.ruff.format]
docstring-code-format = true

[dependency-groups]
dev = [
    "ruff>=0.9.3",
]
